# fix(processor): scale minDetections with detection window duration